refactor: move library manager into dedicated module
